Establishing a new static subdirectory for game
There is probably a way to pass this along as a dynamic variable, but
it's more work than it's worth right now. For now this helps us host the
node service for this in its own subdirectory so that we can have
multiple node based systems on the same web server instance.